Renamed security to authentication and add README documentation
