+++ fsd_utils/__init__.py
@@ -1,6 +1,6 @@
+from fsd_utils import authentication  # noqa
 from fsd_utils import gunicorn  # noqa
 from fsd_utils import logging  # noqa
-from fsd_utils import security  # noqa
 from fsd_utils.config.configclass import configclass  # noqa
 
-__all__ = [configclass, logging, gunicorn, security]
+__all__ = [configclass, logging, gunicorn, authentication]
